Refactor too Many functions on 1 class
